requirements.txt
folium==0.20.0
mcp==1.24.0
openai==2.13.0
python-dotenv==1.2.1
requests==2.32.5
shapely==2.1.2
pyproj

numpy>=2.0
pandas>=2.2

psycopg2-binary==2.9.11

jupyter
ipykernel
